+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 1. [](#improved)
     * Updated `Grav` container object to implement PSR-11 `ContainerInterface`
     * Updated Grav `Processor` classes to implement PSR-15 `MiddlewareInterface`
+    * Make `Data` class to extend `JsonSerializable`
 
 # v1.6.0-beta.4
 ## 10/24/2018